+++ simvue/api/objects/storage/file.py
@@ -1,6 +1,4 @@
-"""
-Simvue File Storage
-===================
+"""Simvue File Storage.
 
 Class for interacting with a file based storage on the server.
 
@@ -9,9 +7,9 @@ Class for interacting with a file based storage on the server.
 import typing
 
 try:
-    from typing import Self
+    from typing import Self, override
 except ImportError:
-    from typing_extensions import Self
+    from typing_extensions import Self, override
 import pydantic
 
 from .base import StorageBase
@@ -19,19 +17,20 @@ from simvue.models import NAME_REGEX
 
 
 class FileStorage(StorageBase):
-    """
-    Simvue File Storage
-    ===================
+    """Simvue File Storage.
 
     This class is used to connect to/create file storage objects on the Simvue server,
     any modification of instance attributes is mirrored on the remote object.
 
     """
 
+    @override
     def __init__(
         self,
         identifier: str | None = None,
-        _read_only: bool = False,
+        *,
+        server_url: str | None = None,
+        server_token: pydantic.SecretStr | None = None,
         **kwargs,
     ) -> None:
         """Initialise a File Storage
@@ -44,11 +43,18 @@ class FileStorage(StorageBase):
         ----------
         identifier : str, optional
             the remote server unique id for the target folder
+        server_url: str | None, optional
+            alternative server URL, default None
+        server_token : str | None, optional
+            token for alternative server, default None
         **kwargs : dict
             any additional arguments to be passed to the object initialiser
         """
-        super().__init__(identifier, **kwargs)
+        super().__init__(
+            identifier, server_url=server_url, server_token=server_token, **kwargs
+        )
 
+    @override
     @classmethod
     @pydantic.validate_call
     def new(
@@ -60,6 +66,8 @@ class FileStorage(StorageBase):
         is_enabled: bool,
         is_default: bool,
         offline: bool = False,
+        server_url: str | None = None,
+        server_token: pydantic.SecretStr | None = None,
         **_,
     ) -> Self:
         """Create a new file storage object.
@@ -78,19 +86,25 @@ class FileStorage(StorageBase):
             if this storage system should become the new default
         offline : bool, optional
             if this instance should be created in offline mode, default False
+        server_url: str | None, optional
+            alternative server URL, default None
+        server_token : str | None, optional
+            token for alternative server, default None
 
         Returns
         -------
         FileStorage
             instance of storage system with staged changes
         """
-        return FileStorage(
+        return cls(
             name=name,
             backend="File",
             disable_check=disable_check,
             is_tenant_useable=is_tenant_useable,
             is_default=is_default,
             is_enabled=is_enabled,
+            server_url=server_url,
+            server_token=server_token,
             _read_only=False,
             _offline=offline,
         )