+++ CHANGELOG.md
@@ -11,12 +11,19 @@
 
 ## [v2.5.2](https://github.com/simvue-io/python-api/releases/tag/v2.5.2) - 2026-05-15
 
+- Fixed issue whereby providing alternative server URL or token arguments was ignored.
 - Fixed legacy event logging support.
 - Added `ca_cert` argument for S3 storages.
 
 ## ~~[v2.5.1](https://github.com/simvue-io/python-api/releases/tag/v2.5.1) - 2026-05-06~~
 
 **Yanked from PyPi**
+
+- ~~Fixed legacy event logging support.~~
+
+## ~~[v2.5.1](https://github.com/simvue-io/python-api/releases/tag/v2.5.1) - 2026-05-06~~
+
+**Yanked from PyPi as issue not fixed.**
 
 - ~~Fixed legacy event logging support.~~